+++ packages/extension-chakra-storefront/src/pages/cart/index.jsx
@@ -4,509 +4,83 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import React, {useState, useMemo, useEffect} from 'react'
-import {FormattedMessage, useIntl} from 'react-intl'
-import {keepPreviousData} from '@tanstack/react-query'
-
-// Chakra Components
-import {Box, Stack, Grid, GridItem, Container, useDisclosure, Button, Flex} from '@chakra-ui/react'
-
-// Project Components
-import CartCta from '../../pages/cart/partials/cart-cta'
-import CartSecondaryButtonGroup from '../../pages/cart/partials/cart-secondary-button-group'
-import CartSkeleton from '../../pages/cart/partials/cart-skeleton'
-import CartTitle from '../../pages/cart/partials/cart-title'
-import ConfirmationModal from '../../components/confirmation-modal'
-import EmptyCart from '../../pages/cart/partials/empty-cart'
-import OrderSummary from '../../components/order-summary'
-import ProductItem from '../../components/product-item'
-import ProductViewModal from '../../components/product-view-modal'
-import BundleProductViewModal from '../../components/product-view-modal/bundle'
-import RecommendedProducts from '../../components/recommended-products'
-
-// Hooks
-import useToast from '../../hooks/use-toast'
-import useNavigation from '../../hooks/use-navigation'
-import {useWishList} from '../../hooks/use-wish-list'
-
-// Constants
-import {
-    API_ERROR_MESSAGE,
-    EINSTEIN_RECOMMENDERS,
-    TOAST_ACTION_VIEW_WISHLIST,
-    TOAST_MESSAGE_ADDED_TO_WISHLIST,
-    TOAST_MESSAGE_REMOVED_ITEM_FROM_CART,
-    TOAST_MESSAGE_ALREADY_IN_WISHLIST
-} from '../../constants'
-import {REMOVE_CART_ITEM_CONFIRMATION_DIALOG_CONFIG} from '../../pages/cart/partials/cart-secondary-button-group'
-
-// Utilities
-import debounce from 'lodash/debounce'
+import React from 'react'
+import {Box, Stack, Grid, GridItem, Container, useDisclosure} from '@chakra-ui/react'
 import {useCurrentBasket, useCurrentCustomer} from '../../hooks/'
-import {
-    useShopperBasketsMutation,
-    useShippingMethodsForShipment,
-    useProducts,
-    useShopperCustomersMutation
-} from '@salesforce/commerce-sdk-react'
-import UnavailableProductConfirmationModal from '../../components/unavailable-product-confirmation-modal'
-import {getUpdateBundleChildArray} from '../../utils/product-utils'
 
-const DEBOUNCE_WAIT = 750
+// Custom Cart Hooks
+import {useCartProducts} from '../../hooks/cart/useCartProducts'
+import {useCartOperations} from '../../hooks/cart/useCartOperations'
+import {useCartWishlist} from '../../hooks/cart/useCartWishlist'
+import {useCartGiftItems} from '../../hooks/cart/useCartGiftItems'
+import {useCartShipping} from '../../hooks/cart/useCartShipping'
+import {useCartErrorHandler} from './utils/cartUtils'
+
+// Cart Components
+import CartTitle from './partials/cart-title'
+import CartProductList from './partials/cart-product-list'
+import CartSummarySection from './partials/cart-summary-section'
+import CartRecommendations from './partials/cart-recommendations'
+import CartModals from './partials/cart-modals'
+import CartSkeleton from './partials/cart-skeleton'
+import EmptyCart from './partials/empty-cart'
+
 const Cart = () => {
     const {data: basket, isPending} = useCurrentBasket()
-    const productIds = basket?.productItems?.map(({productId}) => productId).join(',') ?? ''
-    const {data: products, isPending: isProductsPending} = useProducts(
-        {
-            parameters: {
-                ids: productIds,
-                allImages: true,
-                perPricebook: true
-            }
-        },
-        {
-            enabled: Boolean(productIds),
-            select: (result) => {
-                return result?.data?.reduce((result, item) => {
-                    const key = item.id
-                    result[key] = item
-                    return result
-                }, {})
-            }
-        }
-    )
-
     const {data: customer} = useCurrentCustomer()
-    const {customerId, isRegistered} = customer
+    const {isRegistered} = customer || {}
 
-    /***************** Product Bundles ************************/
-    const bundleChildVariantIds = []
-    basket?.productItems?.forEach((productItem) => {
-        productItem?.bundledProductItems?.forEach((childProduct) => {
-            bundleChildVariantIds.push(childProduct.productId)
-        })
-    })
+    // Error handling
+    const showError = useCartErrorHandler()
 
-    const {data: bundleChildProductData} = useProducts(
-        {
-            parameters: {
-                ids: bundleChildVariantIds?.join(','),
-                allImages: false,
-                expand: ['availability', 'variations'],
-                select: '(data.(id,inventory))'
-            }
-        },
-        {
-            enabled: bundleChildVariantIds?.length > 0,
-            placeholderData: keepPreviousData,
-            select: (result) => {
-                return result?.data?.reduce((result, item) => {
-                    const key = item.id
-                    result[key] = item
-                    return result
-                }, {})
-            }
-        }
+    // Product data and processing
+    const {products, isProductsPending, bundleChildProductData, productsByItemId} =
+        useCartProducts(basket)
+
+    // Cart operations
+    const {
+        selectedItem,
+        setSelectedItem,
+        localQuantity,
+        isCartItemLoading,
+        setCartItemLoading,
+        handleUpdateCart,
+        handleUpdateBundle,
+        handleUnavailableProducts,
+        handleChangeItemQuantity,
+        handleRemoveItem
+    } = useCartOperations(basket, productsByItemId, showError)
+
+    // Wishlist operations
+    const {handleAddToWishlist} = useCartWishlist(showError)
+
+    // Gift items
+    const {localIsGiftItems, handleIsAGiftChange} = useCartGiftItems(
+        basket,
+        setCartItemLoading,
+        setSelectedItem,
+        showError
     )
 
-    // We use the `products` object to reference products by itemId instead of productId
-    // Since with product bundles, even though the parent productId is the same,
-    // variant selection of the bundle children can be different,
-    // and require unique references to each product bundle
-    const productsByItemId = useMemo(() => {
-        const updateProductsByItemId = {}
-        basket?.productItems?.forEach((productItem) => {
-            let currentProduct = products?.[productItem?.productId]
+    // Shipping
+    useCartShipping(basket)
 
-            // calculate inventory for product bundles based on availability of children
-            if (productItem?.bundledProductItems && bundleChildProductData) {
-                let lowestStockLevel =
-                    currentProduct?.inventory?.stockLevel ?? Number.MAX_SAFE_INTEGER
-                let productWithLowestInventory = ''
-                productItem?.bundledProductItems.forEach((bundleChild) => {
-                    const bundleChildStockLevel =
-                        bundleChildProductData?.[bundleChild.productId]?.inventory?.stockLevel ??
-                        Number.MAX_SAFE_INTEGER
-                    lowestStockLevel = Math.min(lowestStockLevel, bundleChildStockLevel)
-                    if (lowestStockLevel === bundleChildStockLevel)
-                        productWithLowestInventory = bundleChild.productName
-                })
-
-                if (currentProduct?.inventory) {
-                    currentProduct = {
-                        ...currentProduct,
-                        inventory: {
-                            ...currentProduct.inventory,
-                            stockLevel: lowestStockLevel,
-                            lowestStockLevelProductName: productWithLowestInventory
-                        }
-                    }
-                }
-            }
-            updateProductsByItemId[productItem.itemId] = currentProduct
-        })
-        return updateProductsByItemId
-    }, [basket, products, bundleChildProductData])
-
-    /*****************Basket Mutation************************/
-    const updateItemInBasketMutation = useShopperBasketsMutation('updateItemInBasket')
-    const updateItemsInBasketMutation = useShopperBasketsMutation('updateItemsInBasket')
-    const removeItemFromBasketMutation = useShopperBasketsMutation('removeItemFromBasket')
-    const updateShippingMethodForShipmentsMutation = useShopperBasketsMutation(
-        'updateShippingMethodForShipment'
-    )
-    /*****************Basket Mutation************************/
-
-    const [selectedItem, setSelectedItem] = useState(undefined)
-    const [localQuantity, setLocalQuantity] = useState({})
-    const [localIsGiftItems, setLocalIsGiftItems] = useState({})
-    const [isCartItemLoading, setCartItemLoading] = useState(false)
-
+    // Modal state
     const {open: isOpen, onOpen, onClose} = useDisclosure()
-    const {formatMessage} = useIntl()
-    const toast = useToast()
-    const navigate = useNavigation()
-    const modalProps = useDisclosure()
 
-    /******************* Shipping Methods for basket shipment *******************/
-    // do this action only if the basket shipping method is not defined
-    // we need to fetch the shippment methods to get the default value before we can add it to the basket
-    const shippingMethodsQuery = useShippingMethodsForShipment(
-        {
-            parameters: {
-                basketId: basket?.basketId,
-                shipmentId: 'me'
-            }
-        },
-        {
-            // only fetch if basket is has no shipping method in the first shipment
-            enabled:
-                !!basket?.basketId &&
-                basket.shipments.length > 0 &&
-                !basket.shipments[0].shippingMethod
-        }
-    )
-
-    // Handle shipping method update when data is returned
-    useEffect(() => {
-        if (!shippingMethodsQuery.isSuccess || !shippingMethodsQuery.data) {
-            return
-        }
-        updateShippingMethodForShipmentsMutation.mutate({
-            parameters: {
-                basketId: basket?.basketId,
-                shipmentId: 'me'
-            },
-            body: {
-                id: shippingMethodsQuery.data.defaultShippingMethodId
-            }
-        })
-    }, [shippingMethodsQuery.isSuccess, shippingMethodsQuery.data])
-
-    /************************* Error handling ***********************/
-    const showError = () => {
-        toast({
-            title: formatMessage(API_ERROR_MESSAGE),
-            type: 'error'
-        })
-    }
-    /************************* Error handling ***********************/
-
-    /**************** Wishlist ****************/
-    const {data: wishlist} = useWishList()
-
-    const createCustomerProductListItem = useShopperCustomersMutation(
-        'createCustomerProductListItem'
-    )
-    const handleAddToWishlist = async (product) => {
-        try {
-            if (!customerId || !wishlist) {
-                return
-            }
-
-            const isItemInWishlist = wishlist?.customerProductListItems?.find(
-                (i) => i.productId === product?.id
-            )
-
-            if (!isItemInWishlist) {
-                await createCustomerProductListItem.mutateAsync({
-                    parameters: {
-                        listId: wishlist.id,
-                        customerId
-                    },
-                    body: {
-                        // NOTE: APi does not respect quantity, it always adds 1
-                        quantity: product.quantity,
-                        productId: product.productId,
-                        public: false,
-                        priority: 1,
-                        type: 'product'
-                    }
-                })
-                toast({
-                    title: formatMessage(TOAST_MESSAGE_ADDED_TO_WISHLIST, {quantity: 1}),
-                    type: 'success',
-                    action: (
-                        <Button variant="link" onClick={() => navigate('/account/wishlist')}>
-                            {formatMessage(TOAST_ACTION_VIEW_WISHLIST)}
-                        </Button>
-                    )
-                })
-            } else {
-                toast({
-                    title: formatMessage(TOAST_MESSAGE_ALREADY_IN_WISHLIST),
-                    type: 'info',
-                    action: (
-                        <Button variant="link" onClick={() => navigate('/account/wishlist')}>
-                            {formatMessage(TOAST_ACTION_VIEW_WISHLIST)}
-                        </Button>
-                    )
-                })
-            }
-        } catch {
-            showError()
-        }
-    }
-    /**************** Wishlist ****************/
-
-    /***************************** Update Cart **************************/
-    const handleUpdateCart = async (variant, quantity) => {
-        // close the modal before handle the change
-        onClose()
-        // using try-catch is better than using onError callback since we have many mutation calls logic here
-        try {
-            setCartItemLoading(true)
-            const productIds = basket.productItems.map(({productId}) => productId)
-
-            // The user is selecting different variant, and it has not existed in basket
-            if (selectedItem.id !== variant.productId && !productIds.includes(variant.productId)) {
-                const item = {
-                    productId: variant.productId,
-                    quantity,
-                    price: variant.price
-                }
-                return await updateItemInBasketMutation.mutateAsync({
-                    parameters: {
-                        basketId: basket.basketId,
-                        itemId: selectedItem.itemId
-                    },
-                    body: item
-                })
-            }
-
-            // The user is selecting different variant, and it has existed in basket
-            // remove this item in the basket, change the quantity for the new selected variant in the basket
-            if (selectedItem.id !== variant.productId && productIds.includes(variant.productId)) {
-                await removeItemFromBasketMutation.mutateAsync({
-                    parameters: {
-                        basketId: basket.basketId,
-                        itemId: selectedItem.itemId
-                    }
-                })
-                const basketItem = basket.productItems.find(
-                    ({productId}) => productId === variant.productId
-                )
-                const newQuantity = quantity + basketItem.quantity
-                return await changeItemQuantity(newQuantity, basketItem)
-            }
-
-            // the user only changes quantity of the same variant
-            if (selectedItem.quantity !== quantity) {
-                return await changeItemQuantity(quantity, selectedItem)
-            }
-        } catch {
-            showError()
-        } finally {
-            setCartItemLoading(false)
-            setSelectedItem(undefined)
-        }
-    }
-
-    const handleUpdateBundle = async (bundle, bundleQuantity, childProducts) => {
-        // close the modal before handle the change
-        onClose()
-
-        try {
-            setCartItemLoading(true)
-            const itemsToBeUpdated = getUpdateBundleChildArray(bundle, childProducts)
-
-            // We only update the parent bundle when the quantity changes
-            // Since top level bundles don't have variants
-            if (bundle.quantity !== bundleQuantity) {
-                itemsToBeUpdated.unshift({
-                    itemId: bundle.itemId,
-                    productId: bundle.productId,
-                    quantity: bundleQuantity
-                })
-            }
-
-            if (itemsToBeUpdated.length) {
-                await updateItemsInBasketMutation.mutateAsync({
-                    method: 'PATCH',
-                    parameters: {
-                        basketId: basket.basketId
-                    },
-                    body: itemsToBeUpdated
-                })
-            }
-        } catch {
-            showError()
-        } finally {
-            setCartItemLoading(false)
-            setSelectedItem(undefined)
-        }
-    }
-
-    const handleIsAGiftChange = async (product, checked) => {
-        try {
-            const previousVal = localIsGiftItems[product.itemId]
-            setLocalIsGiftItems({
-                ...localIsGiftItems,
-                [product.itemId]: checked
-            })
-            setCartItemLoading(true)
-            setSelectedItem(product)
-            await updateItemInBasketMutation.mutateAsync(
-                {
-                    parameters: {basketId: basket?.basketId, itemId: product.itemId},
-                    body: {
-                        productId: product.id,
-                        quantity: parseInt(product.quantity),
-                        gift: checked
-                    }
-                },
-                {
-                    onSettled: () => {
-                        // reset the state
-                        setCartItemLoading(false)
-                        setSelectedItem(undefined)
-                    },
-                    onSuccess: () => {
-                        setLocalIsGiftItems({...localIsGiftItems, [product.itemId]: undefined})
-                    },
-                    onError: () => {
-                        // reset the quantity to the previous value
-                        setLocalIsGiftItems({...localIsGiftItems, [product.itemId]: previousVal})
-                        showError()
-                    }
-                }
-            )
-        } catch (e) {
-            showError()
-        } finally {
-            setCartItemLoading(false)
-            setSelectedItem(undefined)
-        }
-    }
-
-    const handleUnavailableProducts = async (unavailableProductIds) => {
-        const productItems = basket?.productItems?.filter((item) =>
-            unavailableProductIds?.includes(item.productId)
-        )
-
-        await Promise.all(
-            productItems.map(async (item) => {
-                await handleRemoveItem(item)
-            })
-        )
-    }
-    /***************************** Update Cart **************************/
-
-    /***************************** Update quantity **************************/
-    const changeItemQuantity = debounce(async (quantity, product) => {
-        // This local state allows the dropdown to show the desired quantity
-        // while the API call to update it is happening.
-        const previousQuantity = localQuantity[product.itemId]
-        setLocalQuantity({...localQuantity, [product.itemId]: quantity})
-        setCartItemLoading(true)
+    // Handle edit click
+    const handleEditClick = (product) => {
         setSelectedItem(product)
-
-        await updateItemInBasketMutation.mutateAsync(
-            {
-                parameters: {basketId: basket?.basketId, itemId: product.itemId},
-                body: {
-                    productId: product.id,
-                    quantity: parseInt(quantity)
-                }
-            },
-            {
-                onSettled: () => {
-                    // reset the state
-                    setCartItemLoading(false)
-                    setSelectedItem(undefined)
-                },
-                onSuccess: () => {
-                    setLocalQuantity({...localQuantity, [product.itemId]: undefined})
-                },
-                onError: () => {
-                    // reset the quantity to the previous value
-                    setLocalQuantity({...localQuantity, [product.itemId]: previousQuantity})
-                    showError()
-                }
-            }
-        )
-    }, DEBOUNCE_WAIT)
-
-    const handleChangeItemQuantity = async (product, value) => {
-        const stockLevel = productsByItemId?.[product.itemId]?.inventory?.stockLevel ?? 1
-
-        // Handle removing of the items when 0 is selected.
-        if (value === 0) {
-            // Flush last call to keep ui in sync with data.
-            changeItemQuantity.flush()
-
-            // Set the selected item to the current product to the modal acts on it.
-            setSelectedItem(product)
-
-            // Show the modal.
-            modalProps.onOpen()
-
-            // Return false as 0 isn't valid section.
-            return false
-        }
-
-        // Cancel any pending handlers.
-        changeItemQuantity.cancel()
-
-        // Allow use to selected values above the inventory.
-        if (value > stockLevel || value === product.quantity) {
-            return true
-        }
-
-        // Take action.
-        changeItemQuantity(value, product)
-
-        return true
+        onOpen()
     }
-    /***************************** Update quantity **************************/
 
-    /***************************** Remove Item from basket **************************/
-    const handleRemoveItem = async (product) => {
-        setSelectedItem(product)
-        setCartItemLoading(true)
-        await removeItemFromBasketMutation.mutateAsync(
-            {
-                parameters: {basketId: basket.basketId, itemId: product.itemId}
-            },
-            {
-                onSettled: () => {
-                    // reset the state
-                    setCartItemLoading(false)
-                    setSelectedItem(undefined)
-                },
-                onSuccess: () => {
-                    toast({
-                        title: formatMessage(TOAST_MESSAGE_REMOVED_ITEM_FROM_CART, {quantity: 1}),
-                        type: 'success'
-                    })
-                },
-                onError: () => {
-                    showError()
-                }
-            }
-        )
+    // Handle modal close
+    const handleModalClose = () => {
+        onClose()
+        setSelectedItem(undefined)
     }
-    /********* Rendering  UI **********/
+
+    /********* Rendering UI **********/
     if (isPending) {
         return <CartSkeleton />
     }
@@ -514,6 +88,7 @@ const Cart = () => {
     if (!isPending && !basket?.productItems?.length) {
         return <EmptyCart isRegistered={isRegistered} />
     }
+
     return (
         <Box background="gray.50" flex="1" data-testid="sf-cart-container">
             <Container
@@ -531,148 +106,45 @@ const Cart = () => {
                             gap={{base: 10, xl: 20}}
                         >
                             <GridItem>
-                                <Stack gap={4}>
-                                    {basket.productItems?.map((productItem, idx) => {
-                                        return (
-                                            <ProductItem
-                                                key={productItem.itemId}
-                                                index={idx}
-                                                secondaryActions={
-                                                    <CartSecondaryButtonGroup
-                                                        isAGift={
-                                                            localIsGiftItems[productItem.itemId]
-                                                                ? localIsGiftItems[
-                                                                      productItem.itemId
-                                                                  ]
-                                                                : productItem.gift
-                                                        }
-                                                        onIsAGiftChange={handleIsAGiftChange}
-                                                        onAddToWishlistClick={handleAddToWishlist}
-                                                        onEditClick={(product) => {
-                                                            setSelectedItem(product)
-                                                            onOpen()
-                                                        }}
-                                                        onRemoveItemClick={handleRemoveItem}
-                                                    />
-                                                }
-                                                product={{
-                                                    ...productItem,
-                                                    ...(productsByItemId &&
-                                                        productsByItemId[productItem.itemId]),
-                                                    isProductUnavailable: !isProductsPending
-                                                        ? !productsByItemId?.[productItem.itemId]
-                                                        : undefined,
-                                                    price: productItem.price,
-                                                    quantity: localQuantity[productItem.itemId]
-                                                        ? localQuantity[productItem.itemId]
-                                                        : productItem.quantity
-                                                }}
-                                                onItemQuantityChange={handleChangeItemQuantity.bind(
-                                                    this,
-                                                    productItem
-                                                )}
-                                                showLoading={
-                                                    isCartItemLoading &&
-                                                    selectedItem?.itemId === productItem.itemId
-                                                }
-                                                handleRemoveItem={handleRemoveItem}
-                                            />
-                                        )
-                                    })}
-                                </Stack>
-                                <Box>
-                                    {isOpen && !selectedItem.bundledProductItems && (
-                                        <ProductViewModal
-                                            isOpen={isOpen}
-                                            onOpen={onOpen}
-                                            onClose={onClose}
-                                            product={selectedItem}
-                                            updateCart={(variant, quantity) =>
-                                                handleUpdateCart(variant, quantity)
-                                            }
-                                        />
-                                    )}
-                                    {isOpen && selectedItem.bundledProductItems && (
-                                        <BundleProductViewModal
-                                            isOpen={isOpen}
-                                            onOpen={onOpen}
-                                            onClose={onClose}
-                                            product={selectedItem}
-                                            updateCart={(product, quantity, childProducts) =>
-                                                handleUpdateBundle(product, quantity, childProducts)
-                                            }
-                                        />
-                                    )}
-                                </Box>
+                                <CartProductList
+                                    basket={basket}
+                                    productsByItemId={productsByItemId}
+                                    localQuantity={localQuantity}
+                                    localIsGiftItems={localIsGiftItems}
+                                    isProductsPending={isProductsPending}
+                                    isCartItemLoading={isCartItemLoading}
+                                    selectedItem={selectedItem}
+                                    handleChangeItemQuantity={handleChangeItemQuantity}
+                                    handleIsAGiftChange={handleIsAGiftChange}
+                                    handleAddToWishlist={handleAddToWishlist}
+                                    handleEditClick={handleEditClick}
+                                    handleRemoveItem={handleRemoveItem}
+                                />
                             </GridItem>
                             <GridItem>
-                                <Stack gap={4}>
-                                    <OrderSummary
-                                        showPromoCodeForm={true}
-                                        isEstimate={true}
-                                        basket={basket}
-                                    />
-                                    <Box display={{base: 'none', lg: 'block'}}>
-                                        <CartCta />
-                                    </Box>
-                                </Stack>
+                                <CartSummarySection basket={basket} />
                             </GridItem>
                         </Grid>
 
                         {/* Product Recommendations */}
-                        <Stack gap={16}>
-                            <RecommendedProducts
-                                title={
-                                    <FormattedMessage
-                                        defaultMessage="Recently Viewed"
-                                        id="cart.recommended_products.title.recently_viewed"
-                                    />
-                                }
-                                recommender={EINSTEIN_RECOMMENDERS.CART_RECENTLY_VIEWED}
-                                mx={{base: -4, sm: -6, lg: 0}}
-                            />
-
-                            <RecommendedProducts
-                                title={
-                                    <FormattedMessage
-                                        defaultMessage="You May Also Like"
-                                        id="cart.recommended_products.title.may_also_like"
-                                    />
-                                }
-                                recommender={EINSTEIN_RECOMMENDERS.CART_MAY_ALSO_LIKE}
-                                products={basket?.productItems}
-                                shouldFetch={() =>
-                                    basket?.basketId && basket.productItems?.length > 0
-                                }
-                                mx={{base: -4, sm: -6, lg: 0}}
-                            />
-                        </Stack>
+                        <CartRecommendations basket={basket} />
                     </Stack>
                 </Stack>
             </Container>
 
-            <Flex
-                h="130px"
-                position="sticky"
-                bottom={0}
-                bg="white"
-                alignItems="center"
-                flexDirection="column"
-                display={{base: 'flex', lg: 'none'}}
-            >
-                <CartCta />
-            </Flex>
-            <ConfirmationModal
-                {...REMOVE_CART_ITEM_CONFIRMATION_DIALOG_CONFIG}
-                onPrimaryAction={() => {
-                    handleRemoveItem(selectedItem)
-                }}
-                onAlternateAction={() => {}}
-                {...modalProps}
-            />
+            {/* Mobile CTA */}
+            <CartSummarySection basket={basket} isDesktop={false} />
 
-            <UnavailableProductConfirmationModal
-                productItems={basket?.productItems}
+            {/* Modals */}
+            <CartModals
+                isOpen={isOpen}
+                onOpen={onOpen}
+                onClose={handleModalClose}
+                selectedItem={selectedItem}
+                handleUpdateCart={handleUpdateCart}
+                handleUpdateBundle={handleUpdateBundle}
+                handleRemoveItem={handleRemoveItem}
+                basket={basket}
                 handleUnavailableProducts={handleUnavailableProducts}
             />
         </Box>
